# resources/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Evolubat Open Tools",
    "version": "1.3.0",
    "description": "Jeu d'API publiques destinées aux intégrations Open WebUI / OpenIA pour explorer les formations Evolubat."
  },
  "servers": [
    { "url": "http://localhost:8000/api" }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Lister les utilisateurs",
        "operationId": "listUsers",
        "responses": {
          "200": {
            "description": "Liste d'utilisateurs non sensibles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      }
    },
    "/formations": {
      "get": {
        "summary": "Lister toutes les formations",
        "operationId": "listFormations",
        "responses": {
          "200": {
            "description": "Liste de formations brutes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Formation" }
                }
              }
            }
          }
        }
      }
    },
    "/formations/featured": {
      "get": {
        "summary": "Formations mises en avant",
        "operationId": "featuredFormations",
        "responses": {
          "200": {
            "description": "Sélection de formations actives",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "featured": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/FormationHighlight" }
                    },
                    "updated_at": { "type": "string", "format": "date-time" }
                  },
                  "required": ["featured", "updated_at"]
                }
              }
            }
          }
        }
      }
    },
    "/formations/{formation}/outline": {
      "get": {
        "summary": "Structure détaillée d'une formation",
        "operationId": "formationOutline",
        "parameters": [
          {
            "name": "formation",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Identifiant de la formation"
          }
        ],
        "responses": {
          "200": {
            "description": "Outline chapitres/leçons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "formation": { "$ref": "#/components/schemas/FormationSummary" },
                    "outline": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ChapterOutline" }
                    },
                    "refreshed_at": { "type": "string", "format": "date-time" }
                  },
                  "required": ["formation", "outline", "refreshed_at"]
                }
              }
            }
          }
        }
      }
    },
    "/ai/trainers": {
      "get": {
        "summary": "Assistants IA disponibles",
        "operationId": "listAiTrainers",
        "responses": {
          "200": {
            "description": "Assistants actifs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trainers": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AiTrainer" }
                    },
                    "retrieved_at": { "type": "string", "format": "date-time" }
                  },
                  "required": ["trainers", "retrieved_at"]
                }
              }
            }
          }
        }
      }
    },
    "/ai/insights": {
      "get": {
        "summary": "Vue d'ensemble IA + formation",
        "operationId": "aiInsights",
        "responses": {
          "200": {
            "description": "Statistiques agrégées",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AiInsights" }
              }
            }
          }
        }
      }
    },
    "/tools/weather": {
      "get": {
        "summary": "Météo fictive",
        "operationId": "toolWeather",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": { "type": "string", "default": "Paris" },
            "description": "Ville ciblée"
          }
        ],
        "responses": {
          "200": {
            "description": "Observation météo",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WeatherObservation" }
              }
            }
          }
        }
      }
    },
    "/tools/echo": {
      "post": {
        "summary": "Echo JSON utile pour le debug",
        "operationId": "toolEcho",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payload renvoyé avec timestamp",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EchoResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        },
        "required": ["id", "name", "email"]
      },
      "Formation": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "level": { "type": "string" },
          "active": { "type": "boolean" }
        }
      },
      "FormationHighlight": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "level": { "type": "string" },
          "category": { "type": "string", "nullable": true },
          "cover_image_url": { "type": "string", "format": "uri" },
          "learner_count": { "type": "integer" }
        },
        "required": ["id", "title", "description", "level", "cover_image_url", "learner_count"]
      },
      "FormationSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "level": { "type": "string" },
          "category": { "type": "string", "nullable": true },
          "chapter_count": { "type": "integer" },
          "lesson_count": { "type": "integer" }
        },
        "required": ["id", "title", "chapter_count", "lesson_count"]
      },
      "ChapterOutline": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "position": { "type": "integer" },
          "lessons": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LessonOutline" }
          }
        },
        "required": ["id", "title", "lessons"]
      },
      "LessonOutline": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "position": { "type": "integer" }
        },
        "required": ["id", "title"]
      },
      "AiTrainer": {
        "type": "object",
        "properties": {
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "model": { "type": "string" },
          "temperature": { "type": "number", "format": "float" },
          "use_tools": { "type": "boolean" },
          "show_everywhere": { "type": "boolean" },
          "prompt_preview": { "type": "string" },
          "capabilities": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["slug", "name", "model", "temperature", "use_tools", "show_everywhere"]
      },
      "FormationLeaderboard": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "level": { "type": "string", "nullable": true },
          "active": { "type": "boolean" },
          "learner_count": { "type": "integer" },
          "last_update": { "type": "string", "format": "date-time", "nullable": true }
        },
        "required": ["id", "title", "active", "learner_count"]
      },
      "FormationTeaser": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "level": { "type": "string", "nullable": true },
          "summary": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time", "nullable": true }
        },
        "required": ["id", "title", "summary"]
      },
      "AiInsights": {
        "type": "object",
        "properties": {
          "totals": {
            "type": "object",
            "properties": {
              "users": { "type": "integer" },
              "formations": { "type": "integer" },
              "lessons": { "type": "integer" },
              "active_conversations": { "type": "integer" }
            },
            "required": ["users", "formations", "lessons", "active_conversations"]
          },
          "top_formations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FormationLeaderboard" }
          },
          "latest_formations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FormationTeaser" }
          },
          "recent_conversations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "status": { "type": "string" },
                "trainer": { "type": "string", "nullable": true },
                "last_message_at": { "type": "string", "format": "date-time", "nullable": true },
                "user": {
                  "type": "object",
                  "nullable": true,
                  "properties": {
                    "id": { "type": "integer" },
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" }
                  }
                }
              },
              "required": ["id", "status"]
            }
          },
          "recommended_prompts": {
            "type": "array",
            "items": { "type": "string" }
          },
          "generated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["totals", "top_formations", "latest_formations", "recent_conversations", "recommended_prompts", "generated_at"]
      },
      "WeatherObservation": {
        "type": "object",
        "properties": {
          "city": { "type": "string" },
          "temperature_c": { "type": "number", "format": "float" },
          "condition": { "type": "string" },
          "observed_at": { "type": "string", "format": "date-time" }
        },
        "required": ["city", "temperature_c", "condition", "observed_at"]
      },
      "EchoResponse": {
        "type": "object",
        "properties": {
          "received": {
            "type": "object",
            "description": "Payload renvoyé tel qu'envoyé",
            "additionalProperties": true
          },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["received", "timestamp"]
      }
    }
  }
}